requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.0
packaging==23.1
Pillow==10.0.0
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
shiboken6==6.5.2
tqdm==4.66.1
urllib3==2.0.4
yarl==1.9.2
